Move `test`s into `tests` folder
